requirements.txt
python-telegram-bot==20.0b1
googletrans==4.0.0-rc1
indic-transliteration
